fix round

+++ secuencial/base/include/AcceptanceCriterion.hpp
@@ -9,10 +9,10 @@
 using namespace std;
 
 extern std::mt19937 mt;
-extern double temp;
-int metropolisAC1(double costPrevious, double costCurrent);
-int metropolisAC3(double costPrevious, double costCurrent,double Th);
-int dCriteriaAC6(double costPrevious, double costCurrent);
-double p(double costPrevious,double costCurrent);
+extern long double temp;
+int metropolisAC1(long double costPrevious, long double costCurrent);
+int metropolisAC3(long double costPrevious, long double costCurrent,long double Th);
+int dCriteriaAC6(long double costPrevious, long double costCurrent);
+long double p(long double costPrevious,long double costCurrent);
 
 #endif
