+++ secuencial/base/include/ReheatingMethods.hpp
@@ -8,22 +8,22 @@
 using namespace std;
 
 
-void reheatingTR11(double &temp, 
-    double k_reheating,
+void reheatingTR11(long double &temp, 
+    long double k_reheating,
     int n_reheating,
     int count_rechaso);
-void reheatingTR12(double &temp, 
-    double k_reheating,
+void reheatingTR12(long double &temp, 
+    long double k_reheating,
     int n_reheating,
     int count);
-void reheatingTR13(double &temp,
-    double k_reheating,
+void reheatingTR13(long double &temp,
+    long double k_reheating,
     int n_reheating,
     int c_cooling_temperature);
-void reheatingTR14(double &temp,
-    double k_reheating,
+void reheatingTR14(long double &temp,
+    long double k_reheating,
     int k_reheating_init,
     int n_reheating,
     int count_rechaso,
-    double e_const);
+    long double e_const);
 #endif
