+++ secuencial/base/src/AcceptanceCriterion.cpp
@@ -8,14 +8,14 @@
 /// en caso de menor temperatura menor probabibilidad que acepte una solución peor.
 ///////////////////////////////////////////////////
 
-int metropolisAC1(double costPrevious, double costCurrent){
-    uniform_real_distribution<double> dist_accepta(0.0, 1.0);
+int metropolisAC1(long double costPrevious, long double costCurrent){
+    uniform_real_distribution<long double> dist_accepta(0.0, 1.0);
     if(costCurrent < costPrevious){
         return 1;
     }
     else{
-        double valor=p(costPrevious,costCurrent);
-        double nrandom=dist_accepta(mt);
+        long double valor=p(costPrevious,costCurrent);
+        long double nrandom=dist_accepta(mt);
         if(nrandom<valor){
             return 1;
         }
@@ -24,9 +24,9 @@ int metropolisAC1(double costPrevious, double costCurrent){
         }
     }
 }
-double p(double costPrevious,double costCurrent){
-    double po;
-    po = exp(-(costCurrent-costPrevious)/((double)temp));
+long double p(long double costPrevious,long double costCurrent){
+    long double po;
+    po = exp(-(costCurrent-costPrevious)/((long double)temp));
     return po;
 }
 
@@ -35,16 +35,16 @@ double p(double costPrevious,double costCurrent){
 *  
 *
 */
-int metropolisAC3(double costPrevious, double costCurrent,double Th){
-    uniform_real_distribution<double> dist_accepta(0.0, 1.0);
-    //double Th=1.2;
+int metropolisAC3(long double costPrevious, long double costCurrent,long double Th){
+    uniform_real_distribution<long double> dist_accepta(0.0, 1.0);
+    //long double Th=1.2;
     if(costCurrent < costPrevious){
         return 1;
     }
     else{
         if(costPrevious <=costCurrent*Th){
-            double valor=p(costPrevious,costCurrent);
-            double nrandom=dist_accepta(mt);
+            long double valor=p(costPrevious,costCurrent);
+            long double nrandom=dist_accepta(mt);
             if(nrandom<valor){
                 return 1;
             }
@@ -62,7 +62,7 @@ int metropolisAC3(double costPrevious, double costCurrent,double Th){
 /// Deterministic criteria
 ///////////////////////////////////////////////////
 
-int dCriteriaAC6(double costPrevious, double costCurrent){
+int dCriteriaAC6(long double costPrevious, long double costCurrent){
     if(costCurrent - costPrevious <= temp){
         return 1;
     }
